+++ OpenAL32/alFilter.cpp
@@ -41,8 +41,8 @@ void ALfilterState_setParams(ALfilterState *filter, ALfilterType type, float gai
     assert(gain > 0.00001f);
 
     w0 = F_TAU * freq_mult;
-    sin_w0 = sinf(w0);
-    cos_w0 = cosf(w0);
+    sin_w0 = std::sin(w0);
+    cos_w0 = std::cos(w0);
     alpha = sin_w0/2.0f * rcpQ;
 
     /* Calculate filter coefficients depending on filter type */
